pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ai-toolkit"
version = "0.1.0"
description = "Modular Python toolkit for building complex AI toolchains"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
keywords = ["ai", "toolkit", "llama", "gguf", "llm"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "llama-cpp-python>=0.2.76",
  "huggingface_hub>=0.23.0",
  "pydantic>=2.7.0",
]

[project.urls]
Homepage = "https://example.com/ai-toolkit"

[tool.hatch.build]
include = ["ai_toolkit/**", "README.md"]
